More event/state management

+++ app/upload/directives/audio-version.component.ts
@@ -16,7 +16,7 @@ import {AudioModel} from '../models/audio.model';
     </header>
     <section *ngIf="audios">
       <audio-file *ngFor="#audio of audios" [audio]="audio"
-        (cancel)="cancelUpload(audio)"></audio-file>
+        (remove)="removeUpload(audio)"></audio-file>
       <div *ngIf="!audios.length" class="empty">
         <h4>Upload a file to get started</h4>
       </div>
@@ -47,18 +47,24 @@ export class AudioVersionComponent implements OnInit {
         return AudioModel.fromDoc(doc);
       });
       this.uploadService.uploadsForVersion(this.version['id']).forEach((upload) => {
-        this.audios.push(AudioModel.fromUpload(upload, this.version, this.story));
+        this.audios.push(AudioModel.fromUpload(upload, this.version));
       });
     });
   }
 
   addUpload(file: File) {
     let upload = this.uploadService.addFile(this.version['id'], file);
-    this.audios.push(AudioModel.fromUpload(upload, this.version, this.story));
+    this.audios.push(AudioModel.fromUpload(upload, this.version));
   }
 
-  cancelUpload(audio: AudioModel) {
-    this.uploadService.remove(audio.upload);
+  removeUpload(audio: AudioModel) {
+    if (audio.upload) {
+      this.uploadService.remove(audio.upload);
+    }
+    let i = this.audios.indexOf(audio);
+    if (i > -1) {
+      this.audios.splice(i, 1);
+    }
   }
 
 }
